Change InfiltratorGame.Factory to be a nested class
